+++ js/inputmask.js
@@ -928,6 +928,7 @@
 						if (source.match.fn === null && target.match.fn !== null) {
 							return target.match.fn.test(source.match.def, getMaskSet(), pos, false, opts, false);
 						}
+
 						return false;
 					}
 
@@ -1343,7 +1344,7 @@
 
 							if (rslt.refreshFromBuffer) {
 								var refresh = rslt.refreshFromBuffer;
-								strict = true;
+								// strict = true;
 								refreshFromBuffer(refresh === true ? refresh : refresh.start, refresh.end, possibleModifiedBuffer);
 								if (rslt.pos === undefined && rslt.c === undefined) {
 									rslt.pos = getLastValidPosition();
@@ -2938,9 +2939,9 @@
 
 				//apply mask
 				getBufferTemplate(); //initialize the buffer and getmasklength
-				if (el.inputmask._valueGet() !== "" || opts.clearMaskOnLostFocus === false || document.activeElement === el) {
-					var initialValue = $.isFunction(opts.onBeforeMask) ? (opts.onBeforeMask(el.inputmask._valueGet(), opts) || el.inputmask._valueGet()) : el.inputmask._valueGet();
-					if (initialValue !== "") checkVal(el, true, false, initialValue.split(""));
+				if (el.inputmask._valueGet(true) !== "" || opts.clearMaskOnLostFocus === false || document.activeElement === el) {
+					var initialValue = $.isFunction(opts.onBeforeMask) ? (opts.onBeforeMask(el.inputmask._valueGet(true), opts) || el.inputmask._valueGet(true)) : el.inputmask._valueGet(true);
+					if (initialValue !== "") checkVal(el, true, false, isRTL ? initialValue.split("").reverse() : initialValue.split(""));
 					var buffer = getBuffer().slice();
 					undoValue = buffer.join("");
 					// Wrap document.activeElement in a try/catch block since IE9 throw "Unspecified error" if document.activeElement is undefined when we are in an IFrame.